sick checks

+++ downloadDump.sh
@@ -8,12 +8,21 @@ CONFJSON="../allbios.json"
 
 cd $DOWN
 
-rm *md5sums.txt
+checkmd5=$(find . -type f -name "*md5sums.txt"|wc -l)
+#echo ${checkmd5}
+
+if [[ ${checkmd5} -gt 0 ]]; then
+	rm *md5sums.txt
+fi
 wget -c https://dumps.wikimedia.org/wikidatawiki/entities/$DATE/wikidata-$DATE-md5sums.txt -o /dev/null
 
 cat *md5sums.txt | grep 'json.gz' > md5sum.txt
 
-rm *-all.json.gz
+checkjson=$(find . -type f -name "*json.gz"|wc -l)
+
+if [[ ${checkjson} -gt 0 ]]; then
+	rm *-all.json.gz
+fi
 wget -c -t 10 https://dumps.wikimedia.org/wikidatawiki/entities/$DATE/wikidata-$DATE-all.json.gz -o /dev/null
 
 md5sum --strict -c md5sum.txt
